requirements.txt
pyyaml
Dash
dash-bootstrap-components
dash-core-components
pandas
numpy
plotly